backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3
bcrypt==4.0.1
qrcode==7.4.2
pillow==10.0.1
requests==2.31.0
twilio==8.10.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
redis==5.0.1
celery==5.3.4
cryptography==41.0.7
PyJWT==2.8.0